pyproject.toml
[tool.poetry]
name = "transcriptrenamer"
version = "0.1.0"
description = ""
authors = ["ShawnKBeck <48632060+ShawnKBeck@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.2"
python-docx = "^1.1.0"
pyth = "^0.6.0"
pdfkit = "^1.0.0"
striprtf = "^0.0.26"
reportlab = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
